test bash 3

+++ scripts/beforeInstall.bash
@@ -3,4 +3,7 @@ if [ -d /home/ubuntu/build ]; then
 fi
 mkdir -vp /home/ubuntu/build
 
+
+if [[ "$(docker images -q gyim1345/snsapi:latest 2> /dev/null)" !== "" ]]; then
 docker rmi -f $(docker images --format '{{.Repository}}:{{.Tag}}' --filter=reference='gyim1345/snsapi:latest')
+fi
